Add default role to workspace
