+++ packages/twenty-front/src/modules/auth/states/currentWorkspaceState.ts
@@ -1,6 +1,6 @@
 import { createState } from '@ui/utilities/state/utils/createState';
 
-import { Workspace } from '~/generated/graphql';
+import { Role, Workspace } from '~/generated/graphql';
 
 export type CurrentWorkspace = Pick<
   Workspace,
@@ -23,7 +23,9 @@ export type CurrentWorkspace = Pick<
   | 'customDomain'
   | 'workspaceUrls'
   | 'metadataVersion'
->;
+> & {
+  defaultRole: Omit<Role, 'workspaceMembers'> | null;
+};
 
 export const currentWorkspaceState = createState<CurrentWorkspace | null>({
   key: 'currentWorkspaceState',
