+++ packages/twenty-front/src/modules/workspace/graphql/mutations/updateWorkspace.ts
@@ -1,6 +1,8 @@
+import { ROLE_FRAGMENT } from '@/settings/roles/graphql/fragments/roleFragment';
 import { gql } from '@apollo/client';
 
 export const UPDATE_WORKSPACE = gql`
+  ${ROLE_FRAGMENT}
   mutation UpdateWorkspace($input: UpdateWorkspaceInput!) {
     updateWorkspace(data: $input) {
       id
@@ -13,6 +15,9 @@ export const UPDATE_WORKSPACE = gql`
       isGoogleAuthEnabled
       isMicrosoftAuthEnabled
       isPasswordAuthEnabled
+      defaultRole {
+        ...RoleFragment
+      }
     }
   }
 `;
