+++ packages/twenty-server/src/engine/core-modules/workspace/dtos/update-workspace-input.ts
@@ -5,6 +5,7 @@ import {
   IsNotIn,
   IsOptional,
   IsString,
+  IsUUID,
   Matches,
 } from 'class-validator';
 
@@ -184,4 +185,9 @@ export class UpdateWorkspaceInput {
   @IsBoolean()
   @IsOptional()
   isPasswordAuthEnabled?: boolean;
+
+  @Field({ nullable: true })
+  @IsUUID()
+  @IsOptional()
+  defaultRoleId?: string;
 }
