+++ packages/twenty-server/src/engine/metadata-modules/role/role.resolver.ts
@@ -33,23 +33,7 @@ export class RoleResolver {
 
   @Query(() => [RoleDTO])
   async getRoles(@AuthWorkspace() workspace: Workspace): Promise<RoleDTO[]> {
-    const roles = await this.roleService.getWorkspaceRoles(workspace.id);
-
-    return roles.map((role) => ({
-      id: role.id,
-      label: role.label,
-      description: role.description,
-      workspaceId: role.workspaceId,
-      createdAt: role.createdAt,
-      updatedAt: role.updatedAt,
-      isEditable: role.isEditable,
-      userWorkspaceRoles: role.userWorkspaceRoles,
-      canUpdateAllSettings: role.canUpdateAllSettings,
-      canReadAllObjectRecords: role.canReadAllObjectRecords,
-      canUpdateAllObjectRecords: role.canUpdateAllObjectRecords,
-      canSoftDeleteAllObjectRecords: role.canSoftDeleteAllObjectRecords,
-      canDestroyAllObjectRecords: role.canDestroyAllObjectRecords,
-    }));
+    return this.roleService.getWorkspaceRoles(workspace.id);
   }
 
   @Mutation(() => WorkspaceMember)
