+++ packages/twenty-ui/src/display/icon/components/TablerIcons.ts
@@ -12,9 +12,9 @@ export {
   IconArrowDown,
   IconArrowLeft,
   IconArrowRight,
+  IconArrowsVertical,
   IconArrowUp,
   IconArrowUpRight,
-  IconArrowsVertical,
   IconAt,
   IconBaselineDensitySmall,
   IconBell,
@@ -41,8 +41,8 @@ export {
   IconChevronDown,
   IconChevronLeft,
   IconChevronRight,
-  IconChevronUp,
   IconChevronsRight,
+  IconChevronUp,
   IconCircleDot,
   IconCircleOff,
   IconCirclePlus,
@@ -261,6 +261,7 @@ export {
   IconUpload,
   IconUser,
   IconUserCircle,
+  IconUserPin,
   IconUserPlus,
   IconUsers,
   IconVariable,
